backend/requirements.txt
python-snap7
python-dotenv
pytest
fastapi
uvicorn
websockets
asyncio